changes based on review, and cleaning the tests.
