--- gameoflife.test/CustomTestInputData.cs
@@ -1,40 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-
-namespace gameoflife.test
-{
-    public class CustomTestInputData : IEnumerable<object[]>
-    {
-        // todo: auto generate this data.
-        public IEnumerator<object[]> GetEnumerator()
-        {
-            yield return new object[] { @"Generation 1:
-4 8
-........
-....*...
-...**...
-........", 4, 8 };
-            yield return new object[] { @"Generation 1:
-12 12
-............
-........*...
-.......**...
-............
-............
-............
-............
-............
-............
-............
-............
-............", 12, 12 };
-            yield return new object[] { @"Generation 1:
-3 100
-....................................................................................................
-....................................................................................................
-....................................................................................................", 3, 100 };
-        }
-
-        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
-    }
-}
